Add demo files Update usetimer to use setTimeout
